+++ src/models/LocateModel.php
@@ -38,6 +38,32 @@ class LocateModel extends Model
     /** @var array|null */
     public ?array $locationData = null;
 
+    /**
+     * @param array $config
+     */
+    public function __construct(array $config = [])
+    {
+        // Account for a "location" index being set, but not "lat" and "lng", and try to parse that "location" string into a valid latitude and longitude
+        // This makes it possible to paste a comma separated lat/lng string in the field, and have the field return a valid LocateModel even if Google wasn't involved.
+        [
+            'lat' => $lat,
+            'lng' => $lng,
+            'location' => $location,
+        ] = $config + [
+            'lat' => '',
+            'lng' => '',
+            'location' => '',
+        ];
+        if (empty($lat) && empty($lng) && !empty($location) && is_string($location)) {
+            $maybeLatLng = explode(',', preg_replace('/\s+/', '', $location));
+            if (count($maybeLatLng) === 2 && is_numeric($maybeLatLng[0]) && is_numeric($maybeLatLng[1])) {
+                $config['lat'] = $maybeLatLng[0];
+                $config['lng'] = $maybeLatLng[1];
+            }
+        }
+        parent::__construct($config);
+    }
+
     /** @inheritdoc */
     public function rules(): array
     {